+++ src/templates/index.js
@@ -1,4 +1,3 @@
-// const findRecursive = require('../../lib/recursive_match');
 const getName = require('./parsers/_getName');
 const getTemplates = require('./parsers/_getTemplates');
 
@@ -25,19 +24,23 @@ const inlineParsers = Object.assign(
 );
 const bigParsers = Object.assign({}, geo, pronounce, misc, external);
 
+//this gets all the {{template}} strings and decides how to parse them
 const doTemplate = function(tmpl, wiki, r) {
   let name = getName(tmpl);
+
   //we explicitly ignore these templates
   if (ignore.hasOwnProperty(name) === true) {
     wiki = wiki.replace(tmpl, '');
     return wiki;
   }
+
   //string-replacement templates
   if (inlineParsers.hasOwnProperty(name) === true && inlineParsers[name]) {
     let str = inlineParsers[name](tmpl, r);
     wiki = wiki.replace(tmpl, str);
     return wiki;
   }
+
   //section-template parsers
   if (bigParsers.hasOwnProperty(name) === true) {
     let obj = bigParsers[name](tmpl);
@@ -47,6 +50,7 @@ const doTemplate = function(tmpl, wiki, r) {
     wiki = wiki.replace(tmpl, '');
     return wiki;
   }
+
   //fallback parser
   let obj = generic(tmpl, name);
   if (obj) {
@@ -54,6 +58,7 @@ const doTemplate = function(tmpl, wiki, r) {
     wiki = wiki.replace(tmpl, '');
     return wiki;
   }
+
   //bury this template, if we don't know it
   // console.log(`  - no parser for '${name}' -`);
   // console.log('');
@@ -65,12 +70,10 @@ const doTemplate = function(tmpl, wiki, r) {
 //reduce the scary recursive situations
 const allTemplates = function(r, wiki, options) {
   let templates = getTemplates(wiki);
-  // console.log(templates);
   //first, do the nested ones
   templates.nested.forEach(tmpl => {
     wiki = doTemplate(tmpl, wiki, r, options);
   });
-  // console.log(wiki);
   //then, reparse wiki for the top-level ones
   templates = getTemplates(wiki);
   templates.top.forEach(tmpl => {