+++ changelog.md
@@ -67,3 +67,7 @@
 * return a result or undefined for `sentences.bolds(0)`, and the like
 ### 4.2.2
 * support dollar templates
+### 4.5.0
+* support `section(0).wikitext()`
+* support inline {{marriage}} template
+* dangling semi-colons in first-sentence parentheses